Add debounce windowSize hooks to allow for pure JS screensize. Next step: continue making hamburger menu. Todo: find a way to create variables that can be used across js and css
