+++ src/index.js
@@ -3,9 +3,4 @@ import ReactDOM from "react-dom";
 import App from "./App";
 import "./index.scss";
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById("root")
-);
+ReactDOM.render(<App />, document.getElementById("root"));
