+++ src/ubiquitous/components/List/List.js
@@ -2,10 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import "./List.scss";
 
-const List = ({ children }) => {
-  console.log("children are", children);
-  return <ul className="List">{children}</ul>;
-};
+const List = ({ children }) => <ul className="List">{children}</ul>;
 
 export default List;
