--- src/ubiquitous/functions/useDebounce.js
@@ -1,1 +1,0 @@
-export default function useDebounce() {}
